requirements.txt
crewai==0.11.0
langchain==0.1.0
langchain-community==0.0.10
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
protobuf==4.24.4
sqlalchemy==2.0.23
pydantic==2.5.0
typing-extensions==4.8.0 